[ASTRIE] Added a new fortune\n\n*** 305; 12020 H.E.
It doesn't interest me what you do for a living. I want to know what you ache for and if you dare to dream of meeting your heart's longing.

It doesn't interest me how old you are. I want to know if you will risk looking like a fool for love, for your dream, for the adventure of being alive.

— Oriah Mountain Dreamer