[ASTRIE] Added a new fortune: *** 123; 12021 H.E.
 The Hitchhiker's Guide to the Galaxy has a few things to say on
the subject of towels.

 Most importantly, a towel has immense psychological value.  For
some reason, if a non-hitchhiker discovers that a hitchhiker has his towel
with him, he will automatically assume that he is also in possession of a
toothbrush, washcloth, flask, gnat spray, space suit, etc., etc.  Furthermore,
the non-hitchhiker will then happily lend the hitchhiker any of these or
a dozen other items that he may have "lost".  After all, any man who can
hitch the length and breadth of the Galaxy, struggle against terrible odds,
win through and still know where his towel is, is clearly a man to be
reckoned with.

  -- Douglas Adams, "The Hitchhiker's Guide to the Galaxy"