[ASTRIE] Added a new fortune: *** 179; 12021 H.E.
A master programmer passed a novice programmer one day.  The master
noted the novice's preoccupation with a hand-held computer game.  "Excuse me",
he said, "may I examine it?"

The novice bolted to attention and handed the device to the master.
"I see that the device claims to have three levels of play: Easy, Medium,
and Hard", said the master.  "Yet every such device has another level of play,
where the device seeks not to conquer the human, nor to be conquered by the
human."

"Pray, great master," implored the novice, "how does one find this
mysterious setting?"

The master dropped the device to the ground and crushed it under foot.
And suddenly the novice was enlightened.

-- Geoffrey James, "The Tao of Programming"