[ASTRIE] Added a new fortune: ** 317; 12022 H.E.
I was taking a nap tonight around six and somewhere deep down in the dream I was in, it hit me — “I was dreaming.” I felt my unconscious body laying and the mind actively burning calories and sugar giving me this more-vivid-than-life moving picture. I consciously decided to wake myself up by scaring myself with the most profound abstract sensation of fear I could think of. The wake didn’t come. Just for a moment, I was stuck in the nightmare of my own making. In that fleeting blip, I thought to myself, “Is this what Hell is?” The kingdom of self-made fears locked up with a key that was never to be found. It is up to you to make escape but it is never guaranteed.