[ASTRIE] Added a new fortune: ** 338; 12022 H.E.
Someone asked me, "What are humanity's best inventions?" -- probably expecting me to say something around the Internet or computers.

Come to think of it, I can imagine my life without tech and computers -- it wouldn't be much of a problem, quite the opposite. Wouldn't that be nice.

But things that I truly cannot imagine living without, which are essintial to survival include: refrigerator, dish soap, white vinegar, penicilin, and anime.