[ASTRIE] Added a new fortune: ** 330; 12023 H.E.
Those who find ugly meanings in beautiful things are corrupt without being charming. This is a fault. Those who find beautiful meanings in beautiful things are the cultivated. For these there is hope. They are the elect to whom beautiful things mean only Beauty. There is no such thing as a moral or an immoral book. Books are well written, or badly written. That is all.

― Oscar Wilde, /The Picture of Dorian Gray/