[ASTRIE] Added a new fortune: ** 23; 12024 H.E.
MARK MY WORDS...! This drill, will open a hole in the universe. And that hole will be a path for those behind us. The dreams of those who have fallen! The hopes of those who will follow! Those two sets of dreams weave together into a double helix, drilling a path towards tomorrow! And THAT'S Tengen Toppa! That's Gurren-Lagann! My drill is the drill, that creates the HEAVENS!

--- Simon, /Tengen Toppa Gurren Lagann/