[ASTRIE] Added a new fortune: ** 350; 12024 H.E.
We are the victims of an age when men of science are discredited, and only a few remain who are capable of engaging in scientific research. Our philosophers spend all their time in mixing true with false and are interested in nothing but outward show; such little learning as they have they extend on material ends. When they see a man sincere and unremitting in his search for the truth, one who will have nothing to do with falsehood and pretence, they mock and despise him.

--- Omar Khayyam